Big update on laser and ROUGE
ROUGE now contains all the metrics and was adjusted so that it now computes corr in regards to the zscore. Same for laser.
In laser, a depp neural network was employed on the laser embeddings, with extremely good results when compared to the other metrics. With more training, it will be even better.